requirements-dev.txt
# Test dependencies
pytest
pytest-cov
moto[s3,rds]
boto3
requests
python-dateutil
prometheus_client
b2sdk
minio
dropbox
google-api-python-client
google-auth
